--- src/main/java/com/wanted/socialMediaIntegratedFeed/domain/content/ContentRepository.java
@@ -1,4 +1,0 @@
-package com.wanted.socialMediaIntegratedFeed.domain.content;
-
-public interface ContentRepository {
-}
